// src/types/generated/CircuitGraph.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CircuitBlueprintKind } from "./CircuitBlueprintKind";
import type { CircuitEdge } from "./CircuitEdge";
import type { CircuitNode } from "./CircuitNode";

/**
 * The full blueprint AST for one circuit.
 *
 */
export type CircuitGraph = { version: number, blueprint?: CircuitBlueprintKind | null, nodes: Array<CircuitNode>, edges: Array<CircuitEdge>, };